Update directive.ngdoc
There's been a bit of confusion in #angularjs (on irc) about directives what directives are.  I usually reply with something similar to the paragraph I've added.  It's a super-high-level description of directives.  This description seems to clear things up with beginners.  It also has the added benefit of defining "element" and "attribute" with a small, simple, example.  I purposefully wrote it in paragraph form for flow.

This proposed change is inoffensive to dog and cat lovers.  Although dog does come first.  And rightly so.

I also clarified the camelCase to dash-case verbiage a little.